Refactor dankest dungoen meme to generator
